+++ Barik/Utils/ExperimentalConfigurationModifier.swift
@@ -2,14 +2,14 @@ import SwiftUI
 
 private struct ExperimentalConfigurationModifier: ViewModifier {
     @ObservedObject var configManager = ConfigManager.shared
-    var foregroundHeight: CGFloat { configManager.config.experimental.foreground.resolveHeight() }
+    var foregroundHeight: CGFloat { configManager.config.foreground.resolveHeight() }
     
     let horizontalPadding: CGFloat
     let cornerRadius: CGFloat
     
     func body(content: Content) -> some View {
         Group {
-            if !configManager.config.experimental.foreground.widgetsBackground.displayed {
+            if !configManager.config.foreground.widgetsBackground.displayed {
                 content
             } else {
                 content
@@ -17,7 +17,7 @@ private struct ExperimentalConfigurationModifier: ViewModifier {
                     .padding(.horizontal, foregroundHeight < 45 && horizontalPadding != 15 ? 0 :
                                 foregroundHeight < 30 ? 0 : horizontalPadding
                     )
-                    .background(configManager.config.experimental.foreground.widgetsBackground.blur)
+                    .background(configManager.config.foreground.widgetsBackground.blurMaterial)
                     .cornerRadius(foregroundHeight < 30 ? 0 : cornerRadius)
                     .overlay(
                         foregroundHeight < 30 ? nil :